[calendar-events] refactor client code
