+++ packages/read-and-create-calendar-events/client/react/src/styles.js
@@ -1,61 +1,60 @@
 export const styles = {
-  App: {
+  CalendarClient: {
     statusBar: {
-      width: '80vw',
-      margin: '0 auto',
-      height: 60,
-      border: '2px solid #4169e1',
-      background: 'rgba(0,42,245,0.8)',
-      borderRadius: 8,
-      display: 'flex',
       alignContent: 'center',
       alignItems: 'center',
+      background: 'rgba(0,42,245,0.8)',
+      border: '2px solid #4169e1',
+      borderRadius: '8px',
+      display: 'flex',
+      height: '60px',
+      margin: '6em auto 0',
       padding: '8px 16px',
+      width: '80vw',
     },
     statusBarText: {
-      padding: 0,
-      marginBlock: 0,
+      padding: '1em',
+      margin: 0,
       fontWeight: 'bold',
       color: 'white',
     },
     contentContainer: {
+      display: 'flex',
       maxWidth: '80vw',
       margin: '40px auto',
-      display: 'flex',
-      flexDirection: 'row',
-    },
-    agendaContainer: {
-      display: 'flex',
-      flexDirection: 'column',
-      width: '70%',
-      marginRight: 40,
-    },
-    agendaHeader: {
-      display: 'flex',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      marginBottom: 16,
     },
   },
   Agenda: {
     container: {
+      display: 'flex',
+      flexDirection: 'column',
+      width: '70%',
+      marginRight: '40px',
+    },
+    header: {
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      marginBottom: '16px',
+    },
+    eventsContainer: {
       maxHeight: '60vh',
       overflowY: 'scroll',
       border: '4px solid rgb(65, 105, 225)',
-      borderRadius: 16,
-      padding: 12,
+      borderRadius: '16px',
+      padding: '12px',
     },
-    eventArticle: {
+    event: {
       display: 'flex',
       flexDirection: 'column',
       borderRadius: 8,
       boxShadow: '0px 4px 8px rgb(115 115 115 / 60%)',
-      padding: 16,
-      marginBottom: 16,
+      padding: '16px',
+      marginBottom: '16px',
       lineHeight: 1.2,
       whiteSpace: 'pre',
     },
-    eventDateContainer: {
+    eventDate: {
       display: 'flex',
       maxWidth: '50%',
       justifyContent: 'space-between',
@@ -80,19 +79,19 @@ export const styles = {
       flexDirection: 'column',
       width: '30%',
     },
-    title: { marginBottom: 16 },
+    header: { marginBottom: '16px' },
     form: { display: 'flex', flexDirection: 'column' },
-    label: { fontWeight: 'bold', marginBottom: 8 },
+    label: { fontWeight: 'bold', marginBottom: '8px' },
     input: {
-      marginBottom: 16,
+      marginBottom: '16px',
     },
     button: {
       border: 'none',
       outline: 'none',
-      borderRadius: 4,
+      borderRadius: '4px',
       background: '#4169e1',
       color: '#fff',
-      padding: 16,
+      padding: '16px',
       marginBottom: '0.6em',
     },
   },
